# public-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kitchen Master API",
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/dailyPlan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all daily plans for user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daily Plan"
                ],
                "summary": "Get all daily plans for user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.DailyPlan"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a daily plan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daily Plan"
                ],
                "summary": "Create a daily plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "createRecipes",
                        "name": "createRecipes",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.dailyPlanStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/dailyPlan/{dailyPlanId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get specific daily plan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daily Plan"
                ],
                "summary": "Get specific daily plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dailyPlanId",
                        "name": "dailyPlanId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.DailyPlan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a daily plan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daily Plan"
                ],
                "summary": "Update a daily plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dailyPlanId",
                        "name": "dailyPlanId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.dailyPlanStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a daily plan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daily Plan"
                ],
                "summary": "Delete a daily plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dailyPlanId",
                        "name": "dailyPlanId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "required": [
                "content",
                "contentType",
                "status"
            ],
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "contentType": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "api.dailyPlanStoreRequest": {
            "type": "object",
            "required": [
                "day"
            ],
            "properties": {
                "breakfast": {
                    "$ref": "#/definitions/entity.MealPlan"
                },
                "day": {
                    "type": "string"
                },
                "dinner": {
                    "$ref": "#/definitions/entity.MealPlan"
                },
                "lunch": {
                    "$ref": "#/definitions/entity.MealPlan"
                }
            }
        },
...
        "entity.DailyPlan": {
            "type": "object",
            "required": [
                "Id",
                "day",
                "userId"
            ],
            "properties": {
                "Id": {
                    "type": "string"
                },
                "breakfast": {
                    "$ref": "#/definitions/entity.MealPlan"
                },
                "day": {
                    "type": "string"
                },
                "dinner": {
                    "$ref": "#/definitions/entity.MealPlan"
                },
                "lunch": {
                    "$ref": "#/definitions/entity.MealPlan"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "entity.MealPlan": {
            "type": "object",
            "required": [
                "calories",
                "carbohydrates",
                "fats",
                "protein",
                "recipes"
            ],
            "properties": {
                "calories": {
                    "type": "integer"
                },
                "carbohydrates": {
                    "type": "integer"
                },
                "fats": {
                    "type": "integer"
                },
                "protein": {
                    "type": "integer"
                },
                "recipes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Recipe"
                    }
                }
            }
        },
...
        "entity.Recipe": {
            "type": "object",
            "required": [
                "Id",
                "allTime",
                "cookingTime",
                "description",
                "imageUrl",
                "ingredients",
                "instructions",
                "name",
                "notes",
                "portions",
                "preparationTime",
                "tags"
            ],
            "properties": {
                "Id": {
                    "type": "string"
                },
                "allTime": {
                    "type": "string"
                },
                "cookingTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "portions": {
                    "type": "string"
                },
                "preparationTime": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}